1. On windows WId equals void*: - cast WId to qlonglong - add windows-only configure( qlonglong windowId ) overload in order to make DBUS interfaces work
2. Fix usage of KDE_DEPRECATED


svn path=/trunk/KDE/kdepim/akonadi/libakonadi/; revision=773866

+++ agentmanager.cpp
@@ -187,7 +187,7 @@ QString AgentManager::agentInstanceName( const QString &identifier ) const
 
 void AgentManager::agentInstanceConfigure( const QString &identifier, WId windowId )
 {
-  d->mManager->agentInstanceConfigure( identifier, windowId );
+  d->mManager->agentInstanceConfigure( identifier, (qlonglong)windowId );
 }
 
 void AgentManager::agentInstanceSynchronize( const QString &identifier )
